test: move coverage tests out of source files
