+++ src/compiler/translator.rs
@@ -4117,371 +4117,5 @@ impl InstructionTranslator {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    macro_rules! assert_unsupported {
-        ($visitor:ident; $($method:ident($($arg:expr),*)),+ $(,)?) => {
-            $(
-                assert!(matches!(
-                    VisitOperator::$method(&mut $visitor, $($arg),*),
-                    Err(CompilationError::NotSupportedExtension)
-                ));
-            )+
-        };
-    }
-
-    #[test]
-    fn every_unsupported_operator_is_rejected() {
-        let mut visitor =
-            InstructionTranslator::new(Default::default(), false, false, false, false, 0);
-        let memarg = MemArg {
-            align: 0,
-            max_align: 0,
-            offset: 0,
-            memory: 0,
-        };
-        let mut reader = wasmparser::BinaryReader::new(&[
-            0xfd, 0x0c, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-        ]);
-        let v128 = match reader.read_operator().unwrap() {
-            wasmparser::Operator::V128Const { value } => value,
-            operator => panic!("expected v128.const, got {operator:?}"),
-        };
-
-        assert_unsupported!(
-            visitor;
-            visit_try(BlockType::Empty),
-            visit_catch(0),
-            visit_throw(0),
-            visit_rethrow(0),
-            visit_delegate(0),
-            visit_catch_all(),
-            visit_memory_discard(0),
-            visit_memory_atomic_notify(memarg),
-            visit_memory_atomic_wait32(memarg),
-            visit_memory_atomic_wait64(memarg),
-            visit_atomic_fence(),
-            visit_i32_atomic_load(memarg),
-            visit_i64_atomic_load(memarg),
-            visit_i32_atomic_load8_u(memarg),
-            visit_i32_atomic_load16_u(memarg),
-            visit_i64_atomic_load8_u(memarg),
-            visit_i64_atomic_load16_u(memarg),
-            visit_i64_atomic_load32_u(memarg),
-            visit_i32_atomic_store(memarg),
-            visit_i64_atomic_store(memarg),
-            visit_i32_atomic_store8(memarg),
-            visit_i32_atomic_store16(memarg),
-            visit_i64_atomic_store8(memarg),
-            visit_i64_atomic_store16(memarg),
-            visit_i64_atomic_store32(memarg),
-            visit_i32_atomic_rmw_add(memarg),
-            visit_i64_atomic_rmw_add(memarg),
-            visit_i32_atomic_rmw8_add_u(memarg),
-            visit_i32_atomic_rmw16_add_u(memarg),
-            visit_i64_atomic_rmw8_add_u(memarg),
-            visit_i64_atomic_rmw16_add_u(memarg),
-            visit_i64_atomic_rmw32_add_u(memarg),
-            visit_i32_atomic_rmw_sub(memarg),
-            visit_i64_atomic_rmw_sub(memarg),
-            visit_i32_atomic_rmw8_sub_u(memarg),
-            visit_i32_atomic_rmw16_sub_u(memarg),
-            visit_i64_atomic_rmw8_sub_u(memarg),
-            visit_i64_atomic_rmw16_sub_u(memarg),
-            visit_i64_atomic_rmw32_sub_u(memarg),
-            visit_i32_atomic_rmw_and(memarg),
-            visit_i64_atomic_rmw_and(memarg),
-            visit_i32_atomic_rmw8_and_u(memarg),
-            visit_i32_atomic_rmw16_and_u(memarg),
-            visit_i64_atomic_rmw8_and_u(memarg),
-            visit_i64_atomic_rmw16_and_u(memarg),
-            visit_i64_atomic_rmw32_and_u(memarg),
-            visit_i32_atomic_rmw_or(memarg),
-            visit_i64_atomic_rmw_or(memarg),
-            visit_i32_atomic_rmw8_or_u(memarg),
-            visit_i32_atomic_rmw16_or_u(memarg),
-            visit_i64_atomic_rmw8_or_u(memarg),
-            visit_i64_atomic_rmw16_or_u(memarg),
-            visit_i64_atomic_rmw32_or_u(memarg),
-            visit_i32_atomic_rmw_xor(memarg),
-            visit_i64_atomic_rmw_xor(memarg),
-            visit_i32_atomic_rmw8_xor_u(memarg),
-            visit_i32_atomic_rmw16_xor_u(memarg),
-            visit_i64_atomic_rmw8_xor_u(memarg),
-            visit_i64_atomic_rmw16_xor_u(memarg),
-            visit_i64_atomic_rmw32_xor_u(memarg),
-            visit_i32_atomic_rmw_xchg(memarg),
-            visit_i64_atomic_rmw_xchg(memarg),
-            visit_i32_atomic_rmw8_xchg_u(memarg),
-            visit_i32_atomic_rmw16_xchg_u(memarg),
-            visit_i64_atomic_rmw8_xchg_u(memarg),
-            visit_i64_atomic_rmw16_xchg_u(memarg),
-            visit_i64_atomic_rmw32_xchg_u(memarg),
-            visit_i32_atomic_rmw_cmpxchg(memarg),
-            visit_i64_atomic_rmw_cmpxchg(memarg),
-            visit_i32_atomic_rmw8_cmpxchg_u(memarg),
-            visit_i32_atomic_rmw16_cmpxchg_u(memarg),
-            visit_i64_atomic_rmw8_cmpxchg_u(memarg),
-            visit_i64_atomic_rmw16_cmpxchg_u(memarg),
-            visit_i64_atomic_rmw32_cmpxchg_u(memarg),
-            visit_v128_load(memarg),
-            visit_v128_load8x8_s(memarg),
-            visit_v128_load8x8_u(memarg),
-            visit_v128_load16x4_s(memarg),
-            visit_v128_load16x4_u(memarg),
-            visit_v128_load32x2_s(memarg),
-            visit_v128_load32x2_u(memarg),
-            visit_v128_load8_splat(memarg),
-            visit_v128_load16_splat(memarg),
-            visit_v128_load32_splat(memarg),
-            visit_v128_load64_splat(memarg),
-            visit_v128_load32_zero(memarg),
-            visit_v128_load64_zero(memarg),
-            visit_v128_store(memarg),
-            visit_v128_load8_lane(memarg, 0),
-            visit_v128_load16_lane(memarg, 0),
-            visit_v128_load32_lane(memarg, 0),
-            visit_v128_load64_lane(memarg, 0),
-            visit_v128_store8_lane(memarg, 0),
-            visit_v128_store16_lane(memarg, 0),
-            visit_v128_store32_lane(memarg, 0),
-            visit_v128_store64_lane(memarg, 0),
-            visit_v128_const(v128),
-            visit_i8x16_shuffle([0; 16]),
-            visit_i8x16_extract_lane_s(0),
-            visit_i8x16_extract_lane_u(0),
-            visit_i8x16_replace_lane(0),
-            visit_i16x8_extract_lane_s(0),
-            visit_i16x8_extract_lane_u(0),
-            visit_i16x8_replace_lane(0),
-            visit_i32x4_extract_lane(0),
-            visit_i32x4_replace_lane(0),
-            visit_i64x2_extract_lane(0),
-            visit_i64x2_replace_lane(0),
-            visit_f32x4_extract_lane(0),
-            visit_f32x4_replace_lane(0),
-            visit_f64x2_extract_lane(0),
-            visit_f64x2_replace_lane(0),
-            visit_i8x16_swizzle(),
-            visit_i8x16_splat(),
-            visit_i16x8_splat(),
-            visit_i32x4_splat(),
-            visit_i64x2_splat(),
-            visit_f32x4_splat(),
-            visit_f64x2_splat(),
-            visit_i8x16_eq(),
-            visit_i8x16_ne(),
-            visit_i8x16_lt_s(),
-            visit_i8x16_lt_u(),
-            visit_i8x16_gt_s(),
-            visit_i8x16_gt_u(),
-            visit_i8x16_le_s(),
-            visit_i8x16_le_u(),
-            visit_i8x16_ge_s(),
-            visit_i8x16_ge_u(),
-            visit_i16x8_eq(),
-            visit_i16x8_ne(),
-            visit_i16x8_lt_s(),
-            visit_i16x8_lt_u(),
-            visit_i16x8_gt_s(),
-            visit_i16x8_gt_u(),
-            visit_i16x8_le_s(),
-            visit_i16x8_le_u(),
-            visit_i16x8_ge_s(),
-            visit_i16x8_ge_u(),
-            visit_i32x4_eq(),
-            visit_i32x4_ne(),
-            visit_i32x4_lt_s(),
-            visit_i32x4_lt_u(),
-            visit_i32x4_gt_s(),
-            visit_i32x4_gt_u(),
-            visit_i32x4_le_s(),
-            visit_i32x4_le_u(),
-            visit_i32x4_ge_s(),
-            visit_i32x4_ge_u(),
-            visit_i64x2_eq(),
-            visit_i64x2_ne(),
-            visit_i64x2_lt_s(),
-            visit_i64x2_gt_s(),
-            visit_i64x2_le_s(),
-            visit_i64x2_ge_s(),
-            visit_f32x4_eq(),
-            visit_f32x4_ne(),
-            visit_f32x4_lt(),
-            visit_f32x4_gt(),
-            visit_f32x4_le(),
-            visit_f32x4_ge(),
-            visit_f64x2_eq(),
-            visit_f64x2_ne(),
-            visit_f64x2_lt(),
-            visit_f64x2_gt(),
-            visit_f64x2_le(),
-            visit_f64x2_ge(),
-            visit_v128_not(),
-            visit_v128_and(),
-            visit_v128_andnot(),
-            visit_v128_or(),
-            visit_v128_xor(),
-            visit_v128_bitselect(),
-            visit_v128_any_true(),
-            visit_i8x16_abs(),
-            visit_i8x16_neg(),
-            visit_i8x16_popcnt(),
-            visit_i8x16_all_true(),
-            visit_i8x16_bitmask(),
-            visit_i8x16_narrow_i16x8_s(),
-            visit_i8x16_narrow_i16x8_u(),
-            visit_i8x16_shl(),
-            visit_i8x16_shr_s(),
-            visit_i8x16_shr_u(),
-            visit_i8x16_add(),
-            visit_i8x16_add_sat_s(),
-            visit_i8x16_add_sat_u(),
-            visit_i8x16_sub(),
-            visit_i8x16_sub_sat_s(),
-            visit_i8x16_sub_sat_u(),
-            visit_i8x16_min_s(),
-            visit_i8x16_min_u(),
-            visit_i8x16_max_s(),
-            visit_i8x16_max_u(),
-            visit_i8x16_avgr_u(),
-            visit_i16x8_extadd_pairwise_i8x16_s(),
-            visit_i16x8_extadd_pairwise_i8x16_u(),
-            visit_i16x8_abs(),
-            visit_i16x8_neg(),
-            visit_i16x8_q15mulr_sat_s(),
-            visit_i16x8_all_true(),
-            visit_i16x8_bitmask(),
-            visit_i16x8_narrow_i32x4_s(),
-            visit_i16x8_narrow_i32x4_u(),
-            visit_i16x8_extend_low_i8x16_s(),
-            visit_i16x8_extend_high_i8x16_s(),
-            visit_i16x8_extend_low_i8x16_u(),
-            visit_i16x8_extend_high_i8x16_u(),
-            visit_i16x8_shl(),
-            visit_i16x8_shr_s(),
-            visit_i16x8_shr_u(),
-            visit_i16x8_add(),
-            visit_i16x8_add_sat_s(),
-            visit_i16x8_add_sat_u(),
-            visit_i16x8_sub(),
-            visit_i16x8_sub_sat_s(),
-            visit_i16x8_sub_sat_u(),
-            visit_i16x8_mul(),
-            visit_i16x8_min_s(),
-            visit_i16x8_min_u(),
-            visit_i16x8_max_s(),
-            visit_i16x8_max_u(),
-            visit_i16x8_avgr_u(),
-            visit_i16x8_extmul_low_i8x16_s(),
-            visit_i16x8_extmul_high_i8x16_s(),
-            visit_i16x8_extmul_low_i8x16_u(),
-            visit_i16x8_extmul_high_i8x16_u(),
-            visit_i32x4_extadd_pairwise_i16x8_s(),
-            visit_i32x4_extadd_pairwise_i16x8_u(),
-            visit_i32x4_abs(),
-            visit_i32x4_neg(),
-            visit_i32x4_all_true(),
-            visit_i32x4_bitmask(),
-            visit_i32x4_extend_low_i16x8_s(),
-            visit_i32x4_extend_high_i16x8_s(),
-            visit_i32x4_extend_low_i16x8_u(),
-            visit_i32x4_extend_high_i16x8_u(),
-            visit_i32x4_shl(),
-            visit_i32x4_shr_s(),
-            visit_i32x4_shr_u(),
-            visit_i32x4_add(),
-            visit_i32x4_sub(),
-            visit_i32x4_mul(),
-            visit_i32x4_min_s(),
-            visit_i32x4_min_u(),
-            visit_i32x4_max_s(),
-            visit_i32x4_max_u(),
-            visit_i32x4_dot_i16x8_s(),
-            visit_i32x4_extmul_low_i16x8_s(),
-            visit_i32x4_extmul_high_i16x8_s(),
-            visit_i32x4_extmul_low_i16x8_u(),
-            visit_i32x4_extmul_high_i16x8_u(),
-            visit_i64x2_abs(),
-            visit_i64x2_neg(),
-            visit_i64x2_all_true(),
-            visit_i64x2_bitmask(),
-            visit_i64x2_extend_low_i32x4_s(),
-            visit_i64x2_extend_high_i32x4_s(),
-            visit_i64x2_extend_low_i32x4_u(),
-            visit_i64x2_extend_high_i32x4_u(),
-            visit_i64x2_shl(),
-            visit_i64x2_shr_s(),
-            visit_i64x2_shr_u(),
-            visit_i64x2_add(),
-            visit_i64x2_sub(),
-            visit_i64x2_mul(),
-            visit_i64x2_extmul_low_i32x4_s(),
-            visit_i64x2_extmul_high_i32x4_s(),
-            visit_i64x2_extmul_low_i32x4_u(),
-            visit_i64x2_extmul_high_i32x4_u(),
-            visit_f32x4_ceil(),
-            visit_f32x4_floor(),
-            visit_f32x4_trunc(),
-            visit_f32x4_nearest(),
-            visit_f32x4_abs(),
-            visit_f32x4_neg(),
-            visit_f32x4_sqrt(),
-            visit_f32x4_add(),
-            visit_f32x4_sub(),
-            visit_f32x4_mul(),
-            visit_f32x4_div(),
-            visit_f32x4_min(),
-            visit_f32x4_max(),
-            visit_f32x4_pmin(),
-            visit_f32x4_pmax(),
-            visit_f64x2_ceil(),
-            visit_f64x2_floor(),
-            visit_f64x2_trunc(),
-            visit_f64x2_nearest(),
-            visit_f64x2_abs(),
-            visit_f64x2_neg(),
-            visit_f64x2_sqrt(),
-            visit_f64x2_add(),
-            visit_f64x2_sub(),
-            visit_f64x2_mul(),
-            visit_f64x2_div(),
-            visit_f64x2_min(),
-            visit_f64x2_max(),
-            visit_f64x2_pmin(),
-            visit_f64x2_pmax(),
-            visit_i32x4_trunc_sat_f32x4_s(),
-            visit_i32x4_trunc_sat_f32x4_u(),
-            visit_f32x4_convert_i32x4_s(),
-            visit_f32x4_convert_i32x4_u(),
-            visit_i32x4_trunc_sat_f64x2_s_zero(),
-            visit_i32x4_trunc_sat_f64x2_u_zero(),
-            visit_f64x2_convert_low_i32x4_s(),
-            visit_f64x2_convert_low_i32x4_u(),
-            visit_f32x4_demote_f64x2_zero(),
-            visit_f64x2_promote_low_f32x4(),
-            visit_i8x16_relaxed_swizzle(),
-            visit_i32x4_relaxed_trunc_sat_f32x4_s(),
-            visit_i32x4_relaxed_trunc_sat_f32x4_u(),
-            visit_i32x4_relaxed_trunc_sat_f64x2_s_zero(),
-            visit_i32x4_relaxed_trunc_sat_f64x2_u_zero(),
-            visit_f32x4_relaxed_fma(),
-            visit_f32x4_relaxed_fnma(),
-            visit_f64x2_relaxed_fma(),
-            visit_f64x2_relaxed_fnma(),
-            visit_i8x16_relaxed_laneselect(),
-            visit_i16x8_relaxed_laneselect(),
-            visit_i32x4_relaxed_laneselect(),
-            visit_i64x2_relaxed_laneselect(),
-            visit_f32x4_relaxed_min(),
-            visit_f32x4_relaxed_max(),
-            visit_f64x2_relaxed_min(),
-            visit_f64x2_relaxed_max(),
-            visit_i16x8_relaxed_q15mulr_s(),
-            visit_i16x8_dot_i8x16_i7x16_s(),
-            visit_i32x4_dot_i8x16_i7x16_add_s(),
-            visit_f32x4_relaxed_dot_bf16x8_add_f32x4(),
-        );
-    }
-}
+#[path = "../../tests/unit/compiler/translator_tests.rs"]
+mod tests;
